moved to http://forge.octo.com/jcaptcha/confluence
git-svn-id: http://svn.code.sf.net/p/jcaptcha/code/trunk@302 eaf2ec09-4c2e-0410-84d0-fe0bfca4863a

+++ jcaptcha/engine/src/java/com/octo/captcha/engine/image/fisheye/SimpleFishEyeEngine.java
@@ -65,16 +65,13 @@ import java.awt.image.ImageFilter;
  * @author <a href="mailto:mag@jcaptcha.net">Marc-Antoine Garrigue</a>
  * @version 1.0
  */
-public class SimpleFishEyeEngine extends ListImageCaptchaEngine
-{
+public class SimpleFishEyeEngine extends ListImageCaptchaEngine {
 
     /**
-     * this method should be implemented as folow : <ul> <li>First construct all
-     * the factories you want to initialize the gimpy with</li> <li>then call
-     * the this.addFactoriy method for each factory</li> </ul>
+     * this method should be implemented as folow : <ul> <li>First construct all the factories you want to initialize
+     * the gimpy with</li> <li>then call the this.addFactoriy method for each factory</li> </ul>
      */
-    protected void buildInitialFactories()
-    {
+    protected void buildInitialFactories() {
         //build filters
         com.jhlabs.image.SphereFilter sphere = new com.jhlabs.image.SphereFilter();
         com.jhlabs.image.RippleFilter ripple = new com.jhlabs.image.RippleFilter();
@@ -105,23 +102,22 @@ public class SimpleFishEyeEngine extends ListImageCaptchaEngine
         ImageDeformation twirlDef = new ImageDeformationByFilters(
                 new ImageFilter[]{twirl});
 
-
         //add background from files
         BackgroundGenerator generator = new FileReaderRandomBackgroundGenerator(
                 new Integer(250), new Integer(250),
                 "./fisheyebackgrounds");
         addFactory(
-                new FishEyeFactory(generator, sphereDef, new Integer(30),
-                        new Integer(11)));
+                new FishEyeFactory(generator, sphereDef, new Integer(10),
+                        new Integer(5)));
         addFactory(
-                new FishEyeFactory(generator, rippleDef, new Integer(30),
-                        new Integer(11)));
+                new FishEyeFactory(generator, rippleDef, new Integer(10),
+                        new Integer(5)));
         addFactory(
-                new FishEyeFactory(generator, waterDef, new Integer(30),
-                        new Integer(11)));
+                new FishEyeFactory(generator, waterDef, new Integer(10),
+                        new Integer(5)));
         addFactory(
-                new FishEyeFactory(generator, twirlDef, new Integer(30),
-                        new Integer(11)));
+                new FishEyeFactory(generator, twirlDef, new Integer(10),
+                        new Integer(5)));
 
     }
 }
